+++ src/main/java/com/virtukch/dongiveupbe/domain/select_product/dto/SelectProductRequestDto.java
@@ -1,4 +1,27 @@
 package com.virtukch.dongiveupbe.domain.select_product.dto;
 
+import com.virtukch.dongiveupbe.domain.select_product.entity.SelectProduct;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
 public class SelectProductRequestDto {
+    private String selectProductName;
+
+    private String selectProductDescription;
+
+    private Long selectProductViewAmount;
+
+    private String selectProductUrl;
+
+
+    public SelectProduct toEntity() {
+        return SelectProduct.builder()
+                .selectProductName(selectProductName)
+                .selectProductDescription(selectProductDescription)
+                .selectProductViewAmount(selectProductViewAmount)
+                .selectProductUrl(selectProductUrl)
+                .build();
+    }
 }
